+++ packages/br-validators/src/index.ts
@@ -1,1 +1,2 @@
 export * from './cnpj.js';
+export * from './cpf.js';